# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API каталога исполнителей и их альбомов с песнями",
        "description": "API каталога исполнителей и их альбомов с песнями",
        "version": "0.0.1"
    },
    "paths": {
        "/api/albums/": {
            "get": {
                "tags": [
                    "Альбомы"
                ],
                "summary": "Получение информации об альбомах",
                "description": "Получение массива всех альбомов",
                "operationId": "GetAlbums",
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AlbumResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при выполнении запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Альбомы"
                ],
                "summary": "Добавление альбома",
                "description": "Добавление альбома",
                "operationId": "PostAlbum",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AlbumStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AlbumResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "description": "Сообщение ошибки валидатора",
                                            "type": "string",
                                            "example": "Поле 'name' должно быть строкой"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при выполнении запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/albums/{album_id}": {
            "delete": {
                "tags": [
                    "Альбомы"
                ],
                "summary": "Удаление альбома",
                "description": "Удаление альбома и всей информации о нем",
                "operationId": "DeletehAlbum",
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Альбом не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Entity Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при выполнении запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Альбомы"
                ],
                "summary": "Обновление информации об альбоме",
                "description": "Обновление информации об альбоме",
                "operationId": "PatchAlbum",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AlbumUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AlbumResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Альбом не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Entity Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "description": "Сообщение ошибки валидатора",
                                            "type": "string",
                                            "example": "Поле 'name' должно быть строкой"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при выполнении запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/all/": {
            "get": {
                "tags": [
                    "Общая информация"
                ],
                "summary": "Получение всей информации",
                "description": "Получение всех данных",
                "operationId": "GetAllInfo",
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AllExecutorResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при выполнении запроса"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AlbumStoreRequest": {
                "description": "Данные запроса для добавления альбома",
                "required": [
                    "name",
                    "year",
                    "executorId"
                ],
                "properties": {
                    "name": {
                        "description": "Название аольбома",
                        "type": "string",
                        "example": "We Are Not Your Kind"
                    },
                    "year": {
                        "description": "Год выпуска",
                        "type": "integer",
                        "example": "2019"
                    },
                    "executorId": {
                        "description": "Id исполнителя ольбома",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "AlbumUpdateRequest": {
                "description": "Данные запроса для обновления данных альбома",
                "properties": {
                    "name": {
                        "description": "Название аольбома",
                        "type": "string",
                        "example": "We Are Not Your Kind"
                    },
                    "year": {
                        "description": "Год выпуска",
                        "type": "integer",
                        "example": "2019"
                    },
                    "executorId": {
                        "description": "Id исполнителя ольбома",
                        "type": "integer",
                        "example": "2"
                    }
                },
                "type": "object"
            },
...
            "AlbumResource": {
                "description": "Данные об альбоме",
                "properties": {
                    "id": {
                        "description": "Id альбома",
                        "type": "integer",
                        "example": "2"
                    },
                    "name": {
                        "description": "Название альбома",
                        "type": "string",
                        "example": "Iowa"
                    },
                    "yearOfIssue": {
                        "description": "Год выпуска",
                        "type": "integer",
                        "example": "2001"
                    },
                    "executor": {
                        "description": "Информация об исполнителе альбома",
                        "properties": {
                            "id": {
                                "description": "Id исполнителя",
                                "type": "integer",
                                "example": "1"
                            },
                            "name": {
                                "description": "Название исполнителя",
                                "type": "string",
                                "example": "Corey Taylor"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "AllAlbumResource": {
                "description": "Данные об альбоме",
                "properties": {
                    "id": {
                        "description": "Id альбома",
                        "type": "integer",
                        "example": "2"
                    },
                    "name": {
                        "description": "Название альбома",
                        "type": "string",
                        "example": "Iowa"
                    },
                    "yearOfIssue": {
                        "description": "Год выпуска",
                        "type": "integer",
                        "example": "2001"
                    },
                    "songs": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AllSongResource"
                        }
                    }
                },
                "type": "object"
            },
            "AllExecutorResource": {
                "description": "Данные об исполнителях",
                "properties": {
                    "id": {
                        "description": "Id исполнителя",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "Название исполнителя",
                        "type": "string",
                        "example": "Corey Taylor"
                    },
                    "albums": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AllAlbumResource"
                        }
                    }
                },
                "type": "object"
            },
            "AllSongResource": {
                "description": "Данные о песнях",
                "properties": {
                    "id": {
                        "description": "Id песни",
                        "type": "integer",
                        "example": "2"
                    },
                    "name": {
                        "description": "Название песни",
                        "type": "string",
                        "example": "Wait and Bleed"
                    },
                    "orderNumberInAlbum": {
                        "description": "Порядковый номер в альбоме",
                        "type": "integer",
                        "example": "4"
                    }
                },
                "type": "object"
            },
...
        },
...
    }
}